refactor: make internal types pub(crate), move affected tests to unit modules
Address three Copilot review comments on PR #414:

- pub(crate) struct InternalCallChain (src/graph.rs); also pub(crate) on
  find_incoming_chains and find_outgoing_chains since they return this type
- pub(crate) fields prod_chains, test_chains, outgoing_chains on
  FocusedAnalysisOutput (src/analyze.rs); these are #[serde(skip)] and
  serve only internal pagination, not the public API
- pub(crate) fn format_focused_paginated (src/formatter.rs)
- Remove unused dead struct FocusedPaginatedParams entirely

Five integration tests that directly accessed these internal items are moved
into #[cfg(test)] modules inside the respective src/ files so they remain
accessible after the visibility change:
- test_symbol_focus_callers_pagination_first_page  -> src/analyze.rs
- test_symbol_focus_callers_pagination_second_page -> src/analyze.rs
- test_symbol_focus_callees_pagination             -> src/analyze.rs
- test_symbol_focus_empty_prod_callers             -> src/analyze.rs
- test_format_focused_paginated_unit               -> src/formatter.rs
