+++ tests/integration_tests.rs
@@ -3093,185 +3093,6 @@ fn test_file_details_invalid_cursor() {
     assert!(result.is_err(), "invalid cursor should produce an error");
 }
 
-// --- SymbolFocus pagination tests (issue #146) ---
-
-#[test]
-fn test_symbol_focus_callers_pagination_first_page() {
-    use code_analyze_mcp::analyze::analyze_focused;
-    use code_analyze_mcp::pagination::{PaginationMode, paginate_slice};
-
-    let temp_dir = TempDir::new().unwrap();
-
-    // Create a file with many callers of `target`
-    let mut code = String::from("fn target() {}\n");
-    for i in 0..15 {
-        code.push_str(&format!("fn caller_{:02}() {{ target(); }}\n", i));
-    }
-    fs::write(temp_dir.path().join("lib.rs"), &code).unwrap();
-
-    // Act
-    let output = analyze_focused(temp_dir.path(), "target", 1, None, None).unwrap();
-
-    // Paginate prod callers with page_size=5
-    let paginated = paginate_slice(&output.prod_chains, 0, 5, PaginationMode::Callers)
-        .expect("paginate failed");
-    assert!(
-        paginated.total >= 5,
-        "should have enough callers to paginate"
-    );
-    assert!(
-        paginated.next_cursor.is_some(),
-        "should have next_cursor for page 1"
-    );
-
-    // Verify cursor encodes callers mode
-    assert_eq!(paginated.items.len(), 5);
-}
-
-#[test]
-fn test_symbol_focus_callers_pagination_second_page() {
-    use code_analyze_mcp::analyze::analyze_focused;
-    use code_analyze_mcp::formatter::format_focused_paginated;
-    use code_analyze_mcp::pagination::{PaginationMode, decode_cursor, paginate_slice};
-
-    let temp_dir = TempDir::new().unwrap();
-
-    let mut code = String::from("fn target() {}\n");
-    for i in 0..12 {
-        code.push_str(&format!("fn caller_{:02}() {{ target(); }}\n", i));
-    }
-    fs::write(temp_dir.path().join("lib.rs"), &code).unwrap();
-
-    let output = analyze_focused(temp_dir.path(), "target", 1, None, None).unwrap();
-    let total_prod = output.prod_chains.len();
-
-    if total_prod > 5 {
-        // Get page 1 cursor
-        let p1 = paginate_slice(&output.prod_chains, 0, 5, PaginationMode::Callers)
-            .expect("paginate failed");
-        assert!(p1.next_cursor.is_some());
-
-        let cursor_str = p1.next_cursor.unwrap();
-        let cursor_data = decode_cursor(&cursor_str).expect("decode failed");
-
-        // Get page 2
-        let p2 = paginate_slice(
-            &output.prod_chains,
-            cursor_data.offset,
-            5,
-            PaginationMode::Callers,
-        )
-        .expect("paginate failed");
-
-        // Format paginated output
-        let formatted = format_focused_paginated(
-            &p2.items,
-            total_prod,
-            PaginationMode::Callers,
-            "target",
-            &output.prod_chains,
-            &output.test_chains,
-            &output.outgoing_chains,
-            output.def_count,
-            cursor_data.offset,
-            Some(temp_dir.path()),
-            true,
-        );
-
-        // Assert: header shows correct range for page 2
-        let expected_start = cursor_data.offset + 1;
-        assert!(
-            formatted.contains(&format!("CALLERS ({}", expected_start)),
-            "header should show page 2 range, got: {}",
-            formatted
-        );
-    }
-}
-
-#[test]
-fn test_symbol_focus_callees_pagination() {
-    use code_analyze_mcp::analyze::analyze_focused;
-    use code_analyze_mcp::formatter::format_focused_paginated;
-    use code_analyze_mcp::pagination::{PaginationMode, paginate_slice};
-
-    let temp_dir = TempDir::new().unwrap();
-
-    // target calls many functions
-    let mut code = String::from("fn target() {\n");
-    for i in 0..10 {
-        code.push_str(&format!("    callee_{:02}();\n", i));
-    }
-    code.push_str("}\n");
-    for i in 0..10 {
-        code.push_str(&format!("fn callee_{:02}() {{}}\n", i));
-    }
-    fs::write(temp_dir.path().join("lib.rs"), &code).unwrap();
-
-    let output = analyze_focused(temp_dir.path(), "target", 1, None, None).unwrap();
-    let total_callees = output.outgoing_chains.len();
-
-    if total_callees > 3 {
-        let paginated = paginate_slice(&output.outgoing_chains, 0, 3, PaginationMode::Callees)
-            .expect("paginate failed");
-
-        let formatted = format_focused_paginated(
-            &paginated.items,
-            total_callees,
-            PaginationMode::Callees,
-            "target",
-            &output.prod_chains,
-            &output.test_chains,
-            &output.outgoing_chains,
-            output.def_count,
-            0,
-            Some(temp_dir.path()),
-            true,
-        );
-
-        assert!(
-            formatted.contains(&format!(
-                "CALLEES (1-{} of {})",
-                paginated.items.len(),
-                total_callees
-            )),
-            "header should show callees range, got: {}",
-            formatted
-        );
-    }
-}
-
-#[test]
-fn test_symbol_focus_empty_prod_callers() {
-    use code_analyze_mcp::analyze::analyze_focused;
-    use code_analyze_mcp::pagination::{PaginationMode, paginate_slice};
-
-    let temp_dir = TempDir::new().unwrap();
-
-    // target is only called from test functions
-    let code = r#"
-fn target() {}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    #[test]
-    fn test_something() { target(); }
-}
-"#;
-    fs::write(temp_dir.path().join("lib.rs"), code).unwrap();
-
-    let output = analyze_focused(temp_dir.path(), "target", 1, None, None).unwrap();
-
-    // prod_chains may be empty; pagination should handle it gracefully
-    let paginated = paginate_slice(&output.prod_chains, 0, 100, PaginationMode::Callers)
-        .expect("paginate failed");
-    assert_eq!(paginated.items.len(), output.prod_chains.len());
-    assert!(
-        paginated.next_cursor.is_none(),
-        "no next_cursor for empty or single-page prod_chains"
-    );
-}
-
 // --- Unit tests for new formatter functions (issue #146) ---
 
 #[test]
@@ -3351,55 +3172,6 @@ fn test_format_file_details_paginated_unit() {
     assert!(
         !formatted.contains("fn_20"),
         "fn_20 should not be on this page"
-    );
-}
-
-#[test]
-fn test_format_focused_paginated_unit() {
-    use code_analyze_mcp::formatter::format_focused_paginated;
-    use code_analyze_mcp::graph::InternalCallChain;
-    use code_analyze_mcp::pagination::PaginationMode;
-    use std::path::PathBuf;
-
-    // Arrange: create mock caller chains
-    let make_chain = |name: &str| -> InternalCallChain {
-        InternalCallChain {
-            chain: vec![
-                (name.to_string(), PathBuf::from("src/lib.rs"), 10),
-                ("target".to_string(), PathBuf::from("src/lib.rs"), 5),
-            ],
-        }
-    };
-
-    let prod_chains: Vec<InternalCallChain> = (0..8)
-        .map(|i| make_chain(&format!("caller_{}", i)))
-        .collect();
-    let page = &prod_chains[0..3];
-
-    // Act
-    let formatted = format_focused_paginated(
-        page,
-        8,
-        PaginationMode::Callers,
-        "target",
-        &prod_chains,
-        &[],
-        &[],
-        1,
-        0,
-        None,
-        true,
-    );
-
-    // Assert: header present
-    assert!(
-        formatted.contains("CALLERS (1-3 of 8):"),
-        "header should show 1-3 of 8, got: {}",
-        formatted
-    );
-    assert!(
-        formatted.contains("FOCUS: target"),
-        "should have FOCUS header"
     );
 }
